requirements.txt
Django~=5.1.2
django-debug-toolbar~=4.4.6
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.2